Improves performance data laporan page with added icons, new data fetch, and updated layout
The changes in this commit improved the performance data laporan page by adding icons, fetching new data, and updating the layout. The new icons and layout enhance the user experience, making the page more visually appealing and easier to navigate. The updated data fetch ensures that the page displays the most up-to-date information. This change improves the overall performance of the laporan page.

Additionally, the updated layout and new icons align with the overall design principles of the application, maintaining consistency and cohesion throughout the UI. This change also reduces the need for repeated data fetches, improving performance and reducing the load on the server.

By making these improvements, the laporan page now provides a more engaging and informative user experience, allowing users to easily understand and visualize the performance data of the application.
